+++ src/main/java/map/project/musiclibrary/service/ArtistUserService.java
@@ -5,6 +5,7 @@ import map.project.musiclibrary.data.model.audios.Album;
 import map.project.musiclibrary.data.model.audios.Song;
 import map.project.musiclibrary.data.model.users.ArtistUser;
 import map.project.musiclibrary.data.model.users.NormalUser;
+import map.project.musiclibrary.data.model.users.UserSession;
 import map.project.musiclibrary.data.repository.ArtistUserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Lazy;
@@ -30,13 +31,21 @@ public class ArtistUserService {
         this.albumService = albumService;
     }
 
-    public ArtistUser addArtist(String name, String birthdateStr) throws ParseException {
-        SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
-        Date birthdate = dateFormat.parse(birthdateStr);
+    public ArtistUser addArtist(String name, String birthdateStr) {
+        if (UserSession.isLoggedIn() && UserSession.getCurrentUser().isAdmin()) {
+            SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
+            Date birthdate;
+            try {
+                birthdate = dateFormat.parse(birthdateStr);
+            } catch (ParseException e) {
+                throw new IllegalArgumentException("Invalid birthdate format. Please use yyyy-MM-dd.");
+            }
 
-        ArtistUser artist = (ArtistUser) CreatorUserFactory.createCreatorUser("artist", name, birthdate);
+            ArtistUser artist = (ArtistUser) CreatorUserFactory.createCreatorUser("artist", name, birthdate);
 
-        return artistUserRepository.save(artist);
+            return artistUserRepository.save(artist);
+        }
+        throw new SecurityException("Only admin can add an artist");
     }
 
     public ArtistUser save(ArtistUser artistUser) {
@@ -48,7 +57,10 @@ public class ArtistUserService {
     }
 
     public List<ArtistUser> findAll() {
-        return artistUserRepository.findAll();
+        if (UserSession.isLoggedIn() && UserSession.getCurrentUser().isAdmin()) {
+            return artistUserRepository.findAll();
+        }
+        throw new SecurityException("Only admin can list all artists");
     }
 
     public Optional<ArtistUser> findById(Long id) {
@@ -56,40 +68,39 @@ public class ArtistUserService {
     }
 
     public List<NormalUser> getFollowers(String artistIdStr) throws NumberFormatException {
-        Long artistId = Long.parseLong(artistIdStr);
-        Optional<ArtistUser> artistUserOptional = findById(artistId);
-        if (artistUserOptional.isPresent()) {
-            ArtistUser artist = artistUserOptional.get();
-            return artist.getFollowers();
-        } else {
+        if (UserSession.isLoggedIn() && UserSession.getCurrentUser().isAdmin()) {
+            Long artistId = Long.parseLong(artistIdStr);
+            Optional<ArtistUser> artistUserOptional = findById(artistId);
+            if (artistUserOptional.isPresent()) {
+                ArtistUser artist = artistUserOptional.get();
+                return artist.getFollowers();
+            }
             throw new EntityNotFoundException("User with specified id not found");
         }
+        throw new SecurityException("Only admin can list all followers");
     }
 
     @Transactional
     public void delete(String idStr) throws NumberFormatException {
-        Long id = Long.parseLong(idStr);
-        Optional<ArtistUser> optional = artistUserRepository.findById(id);
-        if (optional.isPresent()) {
-            ArtistUser artist = optional.get();
-
-            // remove the link between the artist and label
-            artist.getLabel().getArtists().remove(artist);
-
-            // remove songs associated with the artist
-            // TODO - dependenta circulara fml
-            for (Song song : artist.getSongs()) {
-                songService.delete(song.getId().toString());
+        if (UserSession.isLoggedIn() && UserSession.getCurrentUser().isAdmin()) {
+            Long id = Long.parseLong(idStr);
+            Optional<ArtistUser> optional = artistUserRepository.findById(id);
+            if (optional.isPresent()) {
+                ArtistUser artist = optional.get();
+                // remove the link between the artist and label
+                artist.getLabel().getArtists().remove(artist);
+                // remove songs associated with the artist
+                for (Song song : artist.getSongs()) {
+                    songService.delete(song.getId().toString());
+                }
+                // remove albums associated with the artist
+                for (Album album : artist.getAlbums()) {
+                    albumService.delete(album.getId().toString());
+                }
+                artistUserRepository.deleteById(id);
             }
-
-            // remove albums associated with the artist
-            for (Album album : artist.getAlbums()) {
-                albumService.delete(album.getId().toString());
-            }
-
-            artistUserRepository.deleteById(id);
-        } else {
             throw new EntityNotFoundException("Artist was not found");
         }
+        throw new SecurityException("Only admin can delete an artist");
     }
 }
