+++ src/main/java/map/project/musiclibrary/service/SongService.java
@@ -7,6 +7,7 @@ import map.project.musiclibrary.data.model.strategies.PlayableWithAds;
 import map.project.musiclibrary.data.model.strategies.PlayableWithoutAds;
 import map.project.musiclibrary.data.model.users.ArtistUser;
 import map.project.musiclibrary.data.model.users.NormalUser;
+import map.project.musiclibrary.data.model.users.UserSession;
 import map.project.musiclibrary.data.repository.AdvertisementRepository;
 import map.project.musiclibrary.data.repository.SongRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -32,36 +33,39 @@ public class SongService {
     }
 
     public Song addSong(String name, String genre, String lengthStr, String releaseDateStr, String artistIdStr) throws ParseException {
-        Song song = new Song();
-        song.setName(name);
-        song.setGenre(genre);
+        if (UserSession.isLoggedIn() && UserSession.getCurrentUser().isAdmin()) {
+            Song song = new Song();
+            song.setName(name);
+            song.setGenre(genre);
 
-        try {
-            int length = Integer.parseInt(lengthStr);
-            song.setLength(length);
-        } catch (NumberFormatException e) {
-            throw new NumberFormatException("Invalid integer format. Please provide a valid number.");
-        }
-
-        SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
-        Date releaseDate = dateFormat.parse(releaseDateStr);
-        song.setReleaseDate(releaseDate);
-
-        try {
-            // search artist by id
-            Long artistId = Long.parseLong(artistIdStr);
-            Optional<ArtistUser> artistUserOptional = artistUserService.findById(artistId);
-            if (artistUserOptional.isPresent()) {
-                // add artist to song and add song to artists list
-                song.setArtist(artistUserOptional.get());
-                artistUserOptional.get().addSong(song);
-            } else {
-                throw new IllegalArgumentException("An artist with that id does not exist");
+            try {
+                int length = Integer.parseInt(lengthStr);
+                song.setLength(length);
+            } catch (NumberFormatException e) {
+                throw new NumberFormatException("Invalid integer format. Please provide a valid number.");
             }
-        } catch (NumberFormatException e) {
-            throw new NumberFormatException("Invalid integer format. Please provide a valid number.");
+
+            SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
+            Date releaseDate = dateFormat.parse(releaseDateStr);
+            song.setReleaseDate(releaseDate);
+
+            try {
+                // search artist by id
+                Long artistId = Long.parseLong(artistIdStr);
+                Optional<ArtistUser> artistUserOptional = artistUserService.findById(artistId);
+                if (artistUserOptional.isPresent()) {
+                    // add artist to song and add song to artists list
+                    song.setArtist(artistUserOptional.get());
+                    artistUserOptional.get().addSong(song);
+                } else {
+                    throw new EntityNotFoundException("An artist with that id does not exist");
+                }
+            } catch (NumberFormatException e) {
+                throw new NumberFormatException("Invalid integer format. Please provide a valid number.");
+            }
+            return songRepository.save(song);
         }
-        return songRepository.save(song);
+        throw new SecurityException("Only admin can add a song");
     }
 
     public Song save(Song song) {
@@ -73,36 +77,44 @@ public class SongService {
     }
 
     public List<Song> findAll() {
-        return songRepository.findAll();
+        if (UserSession.isLoggedIn()) {
+            return songRepository.findAll();
+        }
+        throw new SecurityException("You must be logged in to view all songs");
     }
 
     public Optional<Song> findById(Long id) {
         return songRepository.findById(id);
     }
 
-    public String playSong(String songName, NormalUser currentUser) {
-        List<Song> foundSongs = songRepository.findByName(songName);
-        if (!foundSongs.isEmpty()) {
-            return foundSongs.getFirst().play(currentUser.isPremium() ? new PlayableWithoutAds() : new PlayableWithAds(advertisementRepository));
+    public String playSong(String songName) {
+        if (UserSession.isLoggedIn() && UserSession.getCurrentUser().isNormalUser()) {
+            List<Song> foundSongs = songRepository.findByName(songName);
+            if (!foundSongs.isEmpty()) {
+                NormalUser currentUser = (NormalUser) UserSession.getCurrentUser();
+                return foundSongs.getFirst().play(currentUser.isPremium() ? new PlayableWithoutAds() : new PlayableWithAds(advertisementRepository));
+            }
+            throw new EntityNotFoundException("Song was not found");
         }
-        return "Song not found";
+        throw new SecurityException("Only normal users can play songs");
     }
 
     public void delete(String idStr) throws NumberFormatException {
-        Long id = Long.parseLong(idStr);
-        Optional<Song> optional = songRepository.findById(id);
-        if (optional.isPresent()) {
-            Song song = optional.get();
-
-            song.setArtist(null);
-            song.setAlbum(null);
-            for (Playlist playlist : song.getPlaylists()) {
-                playlist.getSongs().remove(song);
+        if (UserSession.isLoggedIn() && UserSession.getCurrentUser().isAdmin()) {
+            Long id = Long.parseLong(idStr);
+            Optional<Song> optional = songRepository.findById(id);
+            if (optional.isPresent()) {
+                Song song = optional.get();
+                song.setArtist(null);
+                song.setAlbum(null);
+                for (Playlist playlist : song.getPlaylists()) {
+                    playlist.getSongs().remove(song);
+                }
+                songRepository.deleteById(id);
+            } else {
+                throw new EntityNotFoundException("Song was not found");
             }
-
-            songRepository.deleteById(id);
-        } else {
-            throw new EntityNotFoundException("Song was not found");
         }
+        throw new SecurityException("Only admin can delete a song");
     }
 }
